Reorganisation of images in repo
