+++ src/assets/animals/chickens/golden_cockerel.gif
