+++ src/assets/animals/chickens/rooster.gif
