+++ src/assets/animals/chickens/undead_chicken.gif
