+++ src/assets/decorations/banners/dawn_breaker_banner.png
