+++ src/assets/decorations/banners/goblin_banner.png
