+++ src/assets/decorations/banners/human_banner.png
