+++ src/assets/decorations/banners/solar_flare_banner.png
