+++ src/assets/sfts/aoe/sir_goldensnout.png
