+++ src/assets/sfts/bears/collectible_bear.png
