+++ src/assets/sfts/bears/cyborg_bear.png
