+++ src/features/game/expansion/components/travel/IslandList.tsx
@@ -17,7 +17,7 @@ import levelUpIcon from "assets/icons/level_up.png";
 import goblin from "assets/buildings/goblin_sign.png";
 import sunflorea from "assets/land/islands/sunflorea.png";
 import snowman from "assets/npcs/snowman.png";
-import dawnBreakerBanner from "assets/decorations/dawn_breaker_banner.png";
+import dawnBreakerBanner from "assets/decorations/banners/dawn_breaker_banner.png";
 import land from "assets/land/islands/island.webp";
 import { SUNNYSIDE } from "assets/sunnyside";
 import { analytics } from "lib/analytics";
