+++ src/features/island/collectibles/components/WitchesEveBanner.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import banner from "assets/decorations/witches_eve_banner.webp";
+import banner from "assets/decorations/banners/witches_eve_banner.webp";
 import { PIXEL_SCALE } from "features/game/lib/constants";
 
 export const WitchesEveBanner: React.FC = () => {
